Make _detach_hidden_room_internal errors non-fatal
When Heisenbridge starts up, it reads join rules for the room and always
treats the first allowed space as the hidden room ID. This assumption is
probably wrong, but the effect is that it conflicts with rooms that were
separately set to allow certain Space members to join.

In such scenarios, Heisenbridge attempts to deconfigure the allowed
spaces on startup, and preventing it from doing so by lowering its
permissions result in an IntentError. Until this is fixed properly,
let's make these errors non-fatal so the bridge can still run with
reduced permissions.